Updates tests. Separates tests into auth, ride, requests
