+++ app/__init__.py
@@ -43,7 +43,7 @@ def create_app(config_name):
             return True
 
     def val_none(**data):
-        '''method to check none value'''
+        '''method to check none'''
         messages = {}
         for key in data:
             if data[key] is None:
@@ -52,12 +52,10 @@ def create_app(config_name):
         return messages
 
     def pass_length(data):
-        """ Checks password length"""
         if len(data) < 8:
             return True
 
     def email_prtn(data):
-        """ Validates user email pattern"""
         pattern = re.match(
             r"(^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z0-9-.]+$)", data)
         if not pattern:
@@ -136,7 +134,7 @@ def create_app(config_name):
 
             if new_user['username'] in all_users:
 
-                return jsonify({'message': 'User already registered'}), 400
+                return jsonify({'message': 'User already registered'}), 404
 
             hashed_pswd = generate_password_hash(new_user['password'])
 
@@ -183,7 +181,7 @@ def create_app(config_name):
         if not request.json:
             abort(404)
         new_ride = {
-            "category": request.json['category'].lower(),
+            "category": request.json['category'],
             "pick_up": request.json['pick_up'],
             "drop_off": request.json['drop_off'],
             "date_time": request.json['date_time'],
@@ -199,7 +197,7 @@ def create_app(config_name):
         categories = [request['category'] for request in rides]
         if new_ride['category'] in categories:
             return jsonify(
-                {'message': 'Failed, Request already made'}), 500
+                {'message': 'Failed, Request already made'}), 404
 
         Rides.create_ride(
             new_ride['category'], new_ride['pick_up'],
@@ -221,7 +219,7 @@ def create_app(config_name):
             if ride[0]['creator_id'] == current_user_id:
                 return jsonify(ride), 200
             return jsonify({'message': 'Not authorized to view ride'})
-        return jsonify({'message': 'Ride not found'}), 404
+        return jsonify({'message': 'Ride not found'}), 200
 
     # Update a specific request
     @app.route("/api/v2/rides/<id>", methods=["PUT"])
@@ -240,13 +238,13 @@ def create_app(config_name):
 
         if category == "" or pick_up == "" or drop_off == "":
 
-            return jsonify({'message': 'Please fill all fields'})
+            return jsonify({'message': 'Please fill all fields'}), 400
 
         message = Rides.update_a_ride(id, category, pick_up, drop_off)
         if message:
-            return jsonify(message), 200
+            return jsonify(message), 201
         else:
-            return ({'message': 'update failed'})
+            return ({'message': 'update failed'}), 400
 
     # Delete a specific request
     @app.route("/api/v2/rides/<id>", methods=["DELETE"])
@@ -259,7 +257,7 @@ def create_app(config_name):
             return jsonify({'message':'Please provide a valid ride Id'}), 400
         ride = Rides.get_a_ride(int(id))
         if len(ride) < 1:
-            return jsonify({'message': 'ride not found'}), 500
+            return jsonify({'message': 'ride not found'}), 400
         else:
             del_id = int(id)
             message = Rides.delete_a_ride(del_id)
@@ -268,7 +266,7 @@ def create_app(config_name):
             if message:
                 return jsonify(message), 200
             else:
-                return jsonify({'message': 'deleting failed'}), 500
+                return jsonify({'message': 'deleting failed'})
 
     @app.route("/api/v2/rides")
     @login_required
@@ -278,10 +276,10 @@ def create_app(config_name):
         print(ride)
         if len(ride) > 0:
             return jsonify(ride), 200
-        return jsonify({'message': 'no rides found'}), 500
+        return jsonify({'message': 'no rides found'}), 400
 
     # ride requests endpoints
-    @app.route("/api/v2/users/rides/<id>/requests", methods=["POST"])
+    @app.route("/api/v2/rides/<id>/requests", methods=["POST"])
     @login_required
     def create_request(current_user_id, id):
         """ Function enables a user to create a request for a ride"""
@@ -308,7 +306,7 @@ def create_app(config_name):
         desc = [request['request_description'] for request in requests]
         if req['request_description'] in desc:
 
-            return jsonify({'message': 'Failed, Request already made'}), 500
+            return jsonify({'message': 'Failed, Request already made'})
 
         Requests.create_request(req['request_description'],
                                 req['request_priority'],
@@ -317,7 +315,7 @@ def create_app(config_name):
         return jsonify({'message': 'Request created successfully'}), 201
 
     # View a specific request
-    @app.route("/api/v2/users/rides/requests/<id>", methods=["GET"])
+    @app.route("/api/v2/rides/requests/<id>", methods=["GET"])
     @login_required
     def get_request(current_user_id, id):
         try:
@@ -330,11 +328,11 @@ def create_app(config_name):
         if request:
             if request[0]['requester_id'] == current_user_id:
                 return jsonify(request), 200
-            return jsonify({'message': 'Not authorized to view request'})
-        return jsonify({'message': 'Request not found'}), 404
+            return jsonify({'message': 'Not authorized to view request'}), 401
+        return jsonify({'message': 'Request not found'})
 
     # Update a specific request
-    @app.route("/api/v2/users/rides/requests/<id>", methods=["PUT"])
+    @app.route("/api/v2/rides/requests/<id>", methods=["PUT"])
     @login_required
     def update_request(current_user_id, id):
         """ Function updated a ride request"""
@@ -349,16 +347,16 @@ def create_app(config_name):
 
         if priority == "" or description == "":
 
-            return jsonify({'message': 'Please fill all fields'})
+            return jsonify({'message': 'Please fill all fields'}), 400
 
         message = Requests.update_a_request(id, description, priority)
         if message:
-            return jsonify(message), 200
+            return jsonify(message), 201
         else:
-            return ({'message': 'update failed'}), 500
+            return ({'message': 'update failed'})
 
     # Delete a specific request
-    @app.route("/api/v2/users/rides/requests/<id>", methods=["DELETE"])
+    @app.route("/api/v2/rides/requests/<id>", methods=["DELETE"])
     @login_required
     def delete_request(current_user_id, id):
         """ Function deletes a ride request from the database"""
@@ -380,7 +378,7 @@ def create_app(config_name):
                 return jsonify({'message': 'deleting failed'})
 
     # User can view all available requests
-    @app.route("/api/v2/users/rides/<id>/requests")
+    @app.route("/api/v2/rides/<id>/requests")
     @login_required
     def get_all_requests(current_user_id, id):
         """ Function gets all ride requests available in the database"""
@@ -388,16 +386,15 @@ def create_app(config_name):
             int(id)
         except ValueError:
             return jsonify({'message':'Please provide a valid request Id'}), 400
-        # request = Requests.get_a_request(int(id))
         request = Requests.get_all_requests(id)
         print(request)
         if len(request) > 0:
             return jsonify(request), 200
 
-        return jsonify({'message': 'no requests found'}), 500
+        return jsonify({'message': 'no requests found'}), 404
 
     # Accept a ride request
-    @app.route("/api/v2/users/rides/requests/<id>/accept", methods=["PUT"])
+    @app.route("/api/v2/rides/requests/<id>/accept", methods=["PUT"])
     @login_required
     def accept_request(current_user_id, id):
         """ Function enables a user to accept a ride request offer"""
@@ -408,16 +405,16 @@ def create_app(config_name):
         status_list = Requests.get_status(id)
         print(status_list)
         if len(status_list) == 0:
-            return jsonify({'message': 'No request found'}), 200
+            return jsonify({'message': 'No request found'}), 404
         status = status_list[0][0].lower()
         if status == "open":
             message = Requests.accept_a_request(id)
             return jsonify(message), 200
         elif status == 'rejected':
-            return jsonify({'message': 'Request already Rejected'}), 500
+            return jsonify({'message': 'Request already Rejected'})
 
     # Reject a request for a ride. Driver action
-    @app.route("/api/v2/users/rides/requests/<id>/reject", methods=["PUT"])
+    @app.route("/api/v2/rides/requests/<id>/reject", methods=["PUT"])
     @login_required
     def reject_request(current_user_id, id):
         """if Users.get_role(current_user_id)[0][0]:
@@ -438,7 +435,7 @@ def create_app(config_name):
             return jsonify(message), 201
         elif status == 'approved':
             return jsonify(
-                {'message': 'Request already accepted'}), 500
+                {'message': 'Request already accepted'})
         # return jsonify({'message': 'Only allowed for the driver'})
